+++ ElvUI_MerathilisUI/Core/Changelog/6.54.lua
@@ -1,8 +1,10 @@
 local MER = unpack(ElvUI_MerathilisUI)
 
 MER.Changelog[654] = {
-	RELEASE_DATE = "TBD",
-	FIXES = {},
+	RELEASE_DATE = "27.02.2025",
+	FIXES = {
+		"[SpecBar]: Fixed an issue which prevent so switch the Specialization- #104",
+	},
 	NEW = {},
 	IMPROVEMENTS = {},
 }