+++ CHANGELOG.md
@@ -1,5 +1,3 @@
 ### Changes ###
 
-  * [11.1]: Changes because of new Patch
-  * [Wowhead Map]: Disable Event Tracker for now, Blizzard got their own one
-  * [AutoButtons]: Add more openable items
+  * [SpecBar]: Fixed an issue which prevent so switch the Specialization- #104